- Experimental Branch Merge #21
git-svn-id: svn://localhost/mosa@923 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/Runtime/CompilerFramework/CIL/BinaryInstruction.cs
@@ -35,7 +35,7 @@ namespace Mosa.Runtime.CompilerFramework.CIL
 		/// </summary>
 		/// <param name="opcode">The opcode.</param>
 		/// <param name="resultCount">The result count.</param>
-		public BinaryInstruction(OpCode opcode, int resultCount)
+		public BinaryInstruction(OpCode opcode, byte resultCount)
 			: base(opcode, 2, resultCount)
 		{
 		}
